Move mock data to forms modules
